feat: 최대 장소 만족시 장소 추가 버튼 비활성화 및 버튼 동기화 해제
